Agregando metodos para eliminar y agregr usuarios

+++ src/main/java/com/dennis/service/UserService.java
@@ -5,9 +5,13 @@ import java.util.List;
 
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.util.ObjectUtils;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.ResponseStatus;
 
 import com.dennis.model.User;
 import com.dennis.repository.UserRepository;
@@ -17,6 +21,10 @@ public class UserService {
 	
 	@Autowired
 	private UserRepository userRepository;
+	
+	public List<User> createUsers(List<User> users) {
+		return userRepository.saveAll(users);
+	}
 	
 	public List<User> getUsers() {
 		return userRepository.findAll();
@@ -50,6 +58,9 @@ public class UserService {
 	
 	public User deleteUserById(String id) {
 		User ActualUser = userRepository.findById(id).orElse(null);
+		if(ObjectUtils.isEmpty(ActualUser) == true) {
+			return ActualUser;
+		}
 		userRepository.delete(ActualUser);
 		return ActualUser;
 	}
